New CreateFlashcards form
This is just the form for creating flashcards. I am trying to make the buttons function, connect the form to the database, but I messed something up. Ya'll should only see the form that doesn't function. Will try to fix my issue with syncing git to visual studio, it keeps giving me warnings.